added more sensor stuff!!!!

+++ main/src/aht30Function/aht30Function.h
@@ -7,7 +7,7 @@ extern bool aht_alive;
 
 typedef struct {
     float temperature;
-    uint8_t humidity;
+    float humidity;
     bool read_flag;
 } AHT_Data_Return;
 AHT_Data_Return readAht30();
